requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
python-dotenv
flask
uvicorn
# fast api
fastapi
requests
reportlab 
pdf2image
Pillow
